Css tweaks to make the style work when we are sometimes forced into light mode
We took the decision to only provide dark mode and turned off the ability for the user to switch to light mode. However, occasionally it seems the browser can override us and set light-mode on the body tag. Till we can find a solution for that and/or clean up all the old light-mode css this workaround makes light-mode look like dark-mode - at least for the main pages.